+++ guizero/PushButton.py
@@ -26,7 +26,7 @@ class PushButton(TextWidget):
         :param Container master:
             The Container (App, Box, etc) the Picture will belong to.
 
-        :param function command:
+        :param Callable command:
             A string containing the image to display, defaults to `None`.
 
         :param List args:
@@ -74,9 +74,6 @@ class PushButton(TextWidget):
             size.
         """
 
-
-        description = "[PushButton] object with text \"" + text + "\""
-
         self._value = 0
         self._image_source = icon
         self._image_source = image
@@ -90,7 +87,7 @@ class PushButton(TextWidget):
         self._text.set(text)
         tk = Button(master.tk, textvariable=self._text, command=self._command_callback)
 
-        super(PushButton, self).__init__(master, tk, description, grid, align, visible, enabled, width, height)
+        super(PushButton, self).__init__(master, tk, grid, align, visible, enabled, width, height)
 
         # Add padding if necessary
         self.tk.config(pady=pady, padx=padx)
@@ -105,6 +102,9 @@ class PushButton(TextWidget):
             self._load_image()
 
     def _load_image(self):
+        # clear the text on the button if an image is being loaded
+        self._text.set("")
+        
         if self._height == "fill" or self._width == "fill":
             utils.raise_error("{}\nCannot use 'fill' for width and height when using a image.".format(self.description))
 
@@ -145,24 +145,27 @@ class PushButton(TextWidget):
     # Set the text on the button
     @text.setter
     def text(self, value):
+        # clear any existing image from the button
+        self._clear_image()
         self._text.set(str(value))
-        self.description = "[Text] object with text \"" + str(value) + "\""
-
+        
     @property
     def image(self):
-        return self._image.image_source
-
+        return None if self._image is None else self._image.image_source
+        
     @image.setter
     def image(self, value):
-        self._image_source = value
-        self._load_image()
+        self._clear_image()
+        if value is not None:
+            self._image_source = value
+            self._load_image()    
 
-
-    def resize(self, width, height):
-        super(PushButton, self.__class__).resize(self, width, height)
-
-        if self._image:
-            self._load_image()
+    @property
+    def description(self):
+        """
+        Returns the description for the widget.
+        """
+        return "[PushButton] object with text '{}'".format(self.text)
 
     # METHODS
     # -------------------------------------------
@@ -192,6 +195,23 @@ class PushButton(TextWidget):
                 self._command = command
             else:
                 self._command = utils.with_args(command, *args)
+    
+    def resize(self, width, height):
+        super(PushButton, self.__class__).resize(self, width, height)
+
+        if self._image:
+            self._load_image()
 
     def _command_callback(self):
         self._command()
+
+    def _clear_image(self):
+        if self._image:
+            self._image_source = None
+            self._image = None
+            self._image_height = None
+            self._image_width = None
+            self._image_player = None
+            self.tk.config(image="")
+            self.resize(None, None)
+            self._text.set(str("Button"))