+++ tests/test_app.py
@@ -1,4 +1,6 @@
-from guizero import App, Text, Picture
+import pytest
+from threading import Event
+from guizero import App, Text, Picture, system_config
 from common_test import (
     schedule_after_test,
     schedule_repeat_test,
@@ -8,7 +10,8 @@ from common_test import (
     cascading_properties_test,
     inheriting_properties_test,
     full_screen_test,
-    add_tk_widget_test)
+    add_tk_widget_test,
+    icon_test)
 
 def test_default_values():
     a = App()
@@ -16,6 +19,7 @@ def test_default_values():
     assert a.width == 500
     assert a.height == 500
     assert a.layout == "auto"
+    assert a.description > ""
     a.destroy()
 
 def test_alt_values():
@@ -70,6 +74,36 @@ def test_events():
     events_test(a)
     a.destroy()
 
+def test_when_resized():
+    
+    a = App()
+
+    resize_event = Event()
+    def callback():
+        resize_event.set()
+
+    def callback_params(event):
+        assert event.width == 503
+        assert event.height == 504
+        resize_event.set()
+
+
+    a.when_resized = callback
+    a.resize(501, 502)
+    assert resize_event.wait(1)
+    resize_event.clear()
+    
+    a.when_resized = callback_params
+    a.resize(503, 504)
+    assert resize_event.wait(1)
+    resize_event.clear()
+
+    a.when_resized = None
+    a.resize(505, 506)
+    assert not resize_event.wait(0.1)
+
+    a.destroy()
+
 def test_cascading_properties():
     a = App()
     cascading_properties_test(a)
@@ -88,4 +122,16 @@ def test_full_screen():
 def test_add_tk_widget():
     a = App()
     add_tk_widget_test(a)
+    a.destroy()
+
+def test_icon():
+    a = App()
+    icon_test(a, "../examples/guizero.gif")
+    a.destroy()
+
+@pytest.mark.skipif(system_config.PIL_available == False,
+                    reason="PIL not available")
+def test_icon_jpg():
+    a = App()
+    icon_test(a, "../examples/guizero.jpg")
     a.destroy()